feat: 기록작성 하단버튼 Modal(#9)
