+++ src/components/UploadModals/ButtonModal.scss
@@ -1,0 +1,66 @@
+@import '../../styles/utils.scss';
+
+
+.buttonModal-wrapper{
+    position: fixed;
+    top:0;
+    left:0;
+    right:0;
+    bottom:0;
+    z-index:15;
+    .dimmed{
+        //position: absolute;
+        width:100%;
+        height:100%;
+        background: #000;
+        opacity: 0.5;
+    }
+    .modal{
+        position:absolute;
+        top:0;
+        left:0;
+        right:0;
+        bottom:0;
+        margin:auto;
+        background-color: white;
+        height:16.2rem;
+        width:30rem;
+        border-radius: 1.2rem;
+        display: flex;
+        flex-direction: column;
+        justify-content: space-between;
+        align-items: center;
+        .text{
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            padding-top:2.2rem;
+            .title{
+                @include font-face(Pretendard-Medium, 1.8rem, 2.6rem, $black90);
+            }
+            .subTitle{
+                @include font-face(Pretendard-Regular, 1.4rem, 1.8rem, $black70);
+                white-space: pre;
+                margin-top:1.2rem;
+            }
+        }
+        .buttons{
+            border-top:0.1rem solid $black40;
+            width:100%;
+            height:4.2rem;
+            button{
+                width:50%;
+                height:100%;
+                padding:0;
+                background-color: transparent;
+                border:none;
+                @include font-face(Pretendard-Semibold, 1.4rem, 1.8rem, $black90);
+            }
+            :first-child{
+                border-right:0.1rem solid $black40;
+                color:$black60;
+            }
+        }
+    }
+    
+}
